Refactor(attendance): Improve schedule interactions and data handling
This commit introduces significant improvements to the attendance feature, focusing on making the `ScheduleScreen` interactive, refining the `UpdateScheduleScreen` logic, and enhancing the underlying data and utility layers.

### Schedule Screen & ViewModel (`ScheduleScreen`, `ScheduleViewModel`)
- **Interactive Attendance:** Users can now directly update the attendance status (Present, Absent, etc.) for a lecture from the `ScheduleScreen`.
- **Repository Integration:** `ScheduleViewModel` now calls `scheduleRepository.upsertLectureSlotAttendanceForDate` to persist attendance changes in the database.
- **State Management:** Added an `isLoading` state to `ScheduleUiState` to provide visual feedback during data operations.
- **Default Date:** The screen now correctly defaults to and loads lectures for the current date on initial launch.

### Update Schedule Screen & ViewModel (`UpdateScheduleScreen`, `UpdateScheduleViewModel`)
- **Form Logic Refinement:**
    - The form now resets correctly when changing days or subjects, preventing stale data.
    - Default values are now set for the selected day (`TimeUtils.currentDayOfWeek()`) and automation segment, improving user experience.
    - Time conflict logic has been refined, preventing overlapping lecture slots.
- **UI Enhancements:**
    - A new `TopLinearLoader` component from `:core:ui` is used to show a progress bar during save/delete operations.
    - Removed the `shouldNavigateBack` logic; navigation is now handled by the navigation graph.
    - The day selector (`DaySelector`) now correctly initializes its scroll position to the current day.
- **ViewModel Logic:**
    - The logic for saving, deleting, and editing lecture slots has been made more robust with proper state updates (`isLoading`, `successMessage`, `errorMessage`).
    - Deleting a lecture now correctly triggers a UI update.
    - Added a comprehensive `resetForm` function to centralize form state management.

### Data & Domain Layer
- **Repository (`ScheduleRepository`, `ScheduleRepositoryImpl`):**
    - Added `upsertLectureSlotAttendanceForDate` to handle inserting or updating attendance records.
    - `deleteLectureSlot` now accepts a `lectureId` instead of a full entity object.
- **DAOs (`AttendanceDao`, `LectureSlotDao`):**
    - `AttendanceDao` now uses Room's `@Upsert` annotation for cleaner code.
    - `LectureSlotDao.deleteLectureSlot` is updated to delete by `lectureId`.
- **Database (`AttendanceDatabase`):**
    - Incremented database version to `4` due to schema changes.
- **Entities:**
    - `LectureSlotEntity`: `locationId` is now nullable.
    - `AttendanceEntity`: `date` type changed from `Long` to `Int`.

### Core & Utility Modules
- **`core/common` (`TimeUtils`, `StringUtils`):**
    - **`TimeUtils`:** Completely refactored for clarity and robustness.
        - Renamed functions for consistency (e.g., `minutesToAmPmString` to `formatMinutesTo12Hour`, `getDayName` to `dayName`).
        - Added new helper functions like `today()`, `currentDayOfWeek()`, and `minutesFromMidnight()`.
    - **`StringUtils`:** The `toCapSmall` and `capitalizeWordLevel` functions are improved with error handling, locale support, and better whitespace trimming.
- **`core/ui`:**
    - Added a new `TopLinearLoader` composable for displaying a linear progress indicator at the top of the screen.
- **`core/analytics`:**
    - Introduced a `CrashlyticsTree` for `Timber` to automatically log errors and exceptions to Firebase Crashlytics in non-debug builds.

### Build & Configuration
- **Dependencies:** Added Firebase Crashlytics dependencies.
- **`minSdk`:** Increased `minSdk` for the `:feature:attendance` module to 26.

+++ feature/attendance/src/main/java/com/reyaz/feature/attendance/data/local/model/AttendanceEntity.kt
@@ -28,8 +28,7 @@ data class AttendanceEntity(
 
     val lectureId: Long,
 
-    // Store as epochDay (LocalDate.toEpochDay())
-    val date: Long,
+    val date: Int,     // Store as epochDay (LocalDate.toEpochDay())
 
     var status: AttendanceStatus
 )
